feat: group ranges
